+++ Source/WebCore/page/RemoteDOMWindow.h
@@ -60,7 +60,6 @@ public:
 
     // DOM API exposed cross-origin.
     WindowProxy* self() const;
-    void close(Document&);
     void focus(LocalDOMWindow& incumbentWindow);
     void blur();
     unsigned length() const;
@@ -68,6 +67,7 @@ public:
     WindowProxy* opener() const;
     void setOpener(WindowProxy*);
     WindowProxy* parent() const;
+    void frameDetached();
     ExceptionOr<void> postMessage(JSC::JSGlobalObject&, LocalDOMWindow& incumbentWindow, JSC::JSValue message, WindowPostMessageOptions&&);
     ExceptionOr<void> postMessage(JSC::JSGlobalObject& globalObject, LocalDOMWindow& incumbentWindow, JSC::JSValue message, String&& targetOrigin, Vector<JSC::Strong<JSC::JSObject>>&& transfer)
     {
@@ -79,6 +79,7 @@ private:
 
     bool isRemoteDOMWindow() const final { return true; }
     bool isLocalDOMWindow() const final { return false; }
+    void closePage() final;
     void setLocation(LocalDOMWindow& activeWindow, const URL& completedURL, SetLocationLocking) final;
 
     WeakPtr<RemoteFrame> m_frame;
