+++ Source/WebCore/page/RemoteFrameClient.h
@@ -47,7 +47,7 @@ public:
     virtual void changeLocation(FrameLoadRequest&&) = 0;
     virtual String renderTreeAsText(size_t baseIndent, OptionSet<RenderAsTextFlag>) = 0;
     virtual void broadcastFrameRemovalToOtherProcesses() = 0;
-    virtual void close() = 0;
+    virtual void closePage() = 0;
     virtual void focus() = 0;
     virtual void unfocus() = 0;
     virtual ~RemoteFrameClient() { }
